# backend/src/main.ts
import Fastify, { type FastifyInstance, type FastifyPluginAsync } from 'fastify';
import cors from '@fastify/cors';
import helmet from '@fastify/helmet';
import rateLimit from '@fastify/rate-limit';
import multipart from '@fastify/multipart';
import swagger from '@fastify/swagger';
import swaggerUi from '@fastify/swagger-ui';
import dotenv from 'dotenv';
import { AppDataSource, assertNoPendingMigrationsForAppStartup } from './config/database.js';
import { errorHandler } from './middleware/errorHandler.js';
import { authRoutes } from './routes/auth.js';
import { businessRoutes } from './routes/businesses.js';
import { dishRoutes } from './routes/dishes.js';
import { menuRoutes } from './routes/menus.js';
import { orderRoutes } from './routes/orders.js';
import { mediaRoutes } from './routes/media.js';
import { reportRoutes } from './routes/reports.js';
import { paymentRoutes } from './routes/payments.js';
import { paymentProcessorRoutes } from './routes/paymentProcessors.js'; // Phase 3: Multi-processor support
import { payoutRoutes } from './routes/payouts.js'; // Phase 3: Automated Payouts
import { taxReportRoutes } from './routes/taxReports.js'; // Phase 3: Tax Compliance & Reporting
import { i18nRoutes } from './routes/i18n.js'; // Phase 3: Multi-Language Support
import { reviewRoutes } from './routes/reviews.js'; // Phase 3: Review & Complaint Workflow
import { marketplaceRoutes } from './routes/marketplace.js'; // Phase 3: Marketplace & Seller Discovery
import { posRoutes } from './routes/pos.js'; // Phase 3: POS System Integration
import { deliveryRoutes } from './routes/delivery.js'; // Phase 3: Delivery Partner Integration
import { couponRoutes } from './routes/coupons.js'; // Phase 3: Promotions & Coupons
import { enhancedReferralRoutes } from './routes/enhancedReferrals.js'; // Phase 3: Enhanced Referral & Viral Features
import { subscriptionRoutes } from './routes/subscriptions.js';
import whatsappRoutes from './routes/whatsapp.js';
import ocrRoutes from './routes/ocr.js';
import referralRoutes from './routes/referrals.js';
import gdprRoutes from './routes/gdpr.js';
import reorderRoutes from './routes/reorder.js';
import adminRoutes from './routes/admin.js';
import { notificationRoutes } from './routes/notifications.js'; // iOS: Notifications
import { cartRoutes } from './routes/cart.js'; // iOS: Shopping cart
import { settingsRoutes } from './routes/settings.js'; // iOS: User settings
import { env } from './config/env.js';
import { requireCapability } from './config/capabilities.js';

// Load environment variables
dotenv.config();

const PORT = env.PORT;
const HOST = process.env.HOST || '0.0.0.0';
const WEBHOOK_RAW_BODY_LIMIT_BYTES = 1024 * 1024;

// Create Fastify instance
const fastify = Fastify({
  logger: {
    level: env.LOG_LEVEL,
    ...(env.NODE_ENV === 'development' && {
      transport: {
        target: 'pino-pretty',
        options: {
          colorize: true,
          translateTime: 'HH:MM:ss Z',
          ignore: 'pid,hostname',
        },
      },
    }),
  },
  // Generate unique request IDs for request tracing
  genReqId: (req) => {
    // Use X-Request-ID header if provided, otherwise generate UUID-style ID
    return req.headers['x-request-id']?.toString() ||
           `req-${Date.now()}-${Math.random().toString(36).substring(2, 11)}`;
  },
  // Disable default request logging (we'll use our custom middleware)
  disableRequestLogging: env.NODE_ENV === 'production',
});

export function shouldCaptureRawWebhookBody(url: string): boolean {
  const path = url.split('?')[0];
  return path === '/api/v1/payments/webhook' || path === '/api/v1/subscriptions/webhook';
}

export function assertWebhookSecretsConfigured(envVars: NodeJS.ProcessEnv = process.env): void {
  if (!['staging', 'production'].includes(envVars.NODE_ENV ?? '')) return;
  const missing = ['STRIPE_WEBHOOK_SECRET', 'STRIPE_WEBHOOK_SECRET_SUBSCRIPTIONS']
    .filter((name) => !envVars[name]);
  if (missing.length) {
    throw new Error(`Missing required webhook signing secret references: ${missing.join(', ')}`);
  }
}

// Register plugins
async function registerPlugins() {
  assertWebhookSecretsConfigured();
  fastify.addContentTypeParser('application/json', {
    parseAs: 'buffer',
    bodyLimit: WEBHOOK_RAW_BODY_LIMIT_BYTES,
  }, (request, body, done) => {
    try {
      if (shouldCaptureRawWebhookBody(request.url)) {
        (request as any).rawBody = Buffer.from(body);
      }
      const text = body.toString('utf8');
      done(null, text.length ? JSON.parse(text) : {});
    } catch (error) {
      done(error as Error, undefined);
    }
  });

  // CORS
  await fastify.register(cors, {
    origin: env.FRONTEND_URL,
    credentials: true,
  });

  // Security headers
  await fastify.register(helmet, {
    contentSecurityPolicy: {
      directives: {
        defaultSrc: ["'self'"],
        styleSrc: ["'self'", "'unsafe-inline'"],
        scriptSrc: ["'self'", "'unsafe-inline'"], // Allow inline for Swagger UI
        imgSrc: ["'self'", 'data:', 'https:'],
      },
    },
  });

  // Swagger/OpenAPI Documentation
  await fastify.register(swagger, {
    openapi: {
      openapi: '3.0.0',
      info: {
        title: 'MenuMaker API',
        description: 'RESTful API for MenuMaker - Restaurant Menu Management & Ordering System',
        version: '1.0.0',
        contact: {
          name: 'MenuMaker Team',
          url: 'https://github.com/ameedanxari/menumaker',
        },
        license: {
          name: 'MIT',
          url: 'https://opensource.org/licenses/MIT',
        },
      },
      servers: [
        {
          url: 'http://localhost:3001',
          description: 'Development server',
        },
        {
          url: 'https://api.menumaker.app',
          description: 'Production server',
        },
      ],
      tags: [
        { name: 'auth', description: 'Authentication endpoints' },
        { name: 'businesses', description: 'Business management endpoints' },
        { name: 'dishes', description: 'Dish and category management endpoints' },
        { name: 'menus', description: 'Menu management endpoints' },
        { name: 'orders', description: 'Order management endpoints' },
        { name: 'media', description: 'File upload and management endpoints' },
        { name: 'reports', description: 'Reporting and analytics endpoints' },
      ],
      components: {
        securitySchemes: {
          bearerAuth: {
            type: 'http',
            scheme: 'bearer',
            bearerFormat: 'JWT',
            description: 'Enter your JWT token obtained from the login endpoint',
          },
        },
      },
    },
  });

  await fastify.register(swaggerUi, {
    routePrefix: '/api/docs',
    uiConfig: {
      docExpansion: 'list',
      deepLinking: true,
      displayRequestDuration: true,
      filter: true,
      showExtensions: true,
      showCommonExtensions: true,
    },
    staticCSP: true,
    transformStaticCSP: (header) => header,
  });

  // Rate limiting
  await fastify.register(rateLimit, {
    max: parseInt(process.env.RATE_LIMIT_MAX || '100', 10),
    timeWindow: parseInt(process.env.RATE_LIMIT_WINDOW_MS || '900000', 10), // 15 minutes
  });

  // Multipart/form-data support (for file uploads)
  await fastify.register(multipart, {
    limits: {
      fileSize: 5 * 1024 * 1024, // 5MB
      files: 10,
    },
  });
}

interface CapabilityGateOptions {
  bypassPaths?: string[];
}

export function shouldBypassCapabilityGate(
  requestUrl: string,
  gatedPrefix: string,
  bypassPaths: string[] = []
): boolean {
  const path = requestUrl.split('?')[0];
  return bypassPaths.some((bypassPath) => {
    const normalizedBypass = bypassPath.startsWith('/') ? bypassPath : `/${bypassPath}`;
    return path === normalizedBypass || path === `${gatedPrefix}${normalizedBypass}`;
  });
}

async function registerCapabilityGatedRoutes(
  app: FastifyInstance,
  routes: FastifyPluginAsync,
  prefix: string,
  capability: string,
  options: CapabilityGateOptions = {}
): Promise<void> {
  await app.register(async (scoped) => {
    const capabilityGate = requireCapability(capability);
    scoped.addHook('preHandler', async (request, reply) => {
      if (shouldBypassCapabilityGate(request.url, prefix, options.bypassPaths)) {
        return;
      }
      return capabilityGate(request, reply);
    });
    await scoped.register(routes);
  }, { prefix });
}

// Register routes
async function registerRoutes() {
  // Global request logging hook
  fastify.addHook('onRequest', async (request, reply) => {
    request.log.info({
      msg: 'Incoming request',
      requestId: request.id,
      method: request.method,
      url: request.url,
      userAgent: request.headers['user-agent'],
      ip: request.ip,
    });

    // Track request start time for duration calculation
    (request as any).startTime = Date.now();

    // Add response hook to log completion
    // @ts-expect-error - addHook on reply is non-standard but works
    reply.addHook('onSend', async () => {
      const duration = Date.now() - (request as any).startTime;
      const isProduction = process.env.NODE_ENV === 'production';
      const isSlow = duration > 1000;
      const isError = reply.statusCode >= 400;

      // Log all requests in dev, only slow or error requests in production
      if (!isProduction || isSlow || isError) {
        request.log.info({
          msg: 'Request completed',
          requestId: request.id,
          method: request.method,
          url: request.url,
          statusCode: reply.statusCode,
          duration,
          ...(isSlow && { performance: 'slow' }),
          ...(isError && { level: 'warn' }),
        });
      }
    });
  });

  // Health check
  fastify.get('/health', async () => {
    return {
      status: 'ok',
      timestamp: new Date().toISOString(),
      uptime: process.uptime(),
    };
  });

  // API routes
  await fastify.register(authRoutes, { prefix: '/api/v1/auth' });
  await fastify.register(businessRoutes, { prefix: '/api/v1/businesses' });
  await fastify.register(dishRoutes, { prefix: '/api/v1/dishes' });
  await fastify.register(menuRoutes, { prefix: '/api/v1/menus' });
  await fastify.register(orderRoutes, { prefix: '/api/v1/orders' });
  await fastify.register(mediaRoutes, { prefix: '/api/v1/media' });
  await fastify.register(reportRoutes, { prefix: '/api/v1/reports' });
  await fastify.register(paymentRoutes, { prefix: '/api/v1/payments' });
  await fastify.register(paymentProcessorRoutes, { prefix: '/api/v1/payment-processors' }); // Phase 3: Multi-processor support
  await fastify.register(payoutRoutes, { prefix: '/api/v1/payouts' }); // Phase 3: Automated Payouts
  await registerCapabilityGatedRoutes(fastify, taxReportRoutes, '/api/v1/tax', 'tax_reporting'); // Disabled until tax reporting is launched.
  await fastify.register(i18nRoutes, { prefix: '/api/v1/i18n' }); // Phase 3: Multi-Language Support
  await fastify.register(reviewRoutes, { prefix: '/api/v1/reviews' }); // Phase 3: Review & Complaint Workflow
  await fastify.register(marketplaceRoutes, { prefix: '/api/v1/marketplace' }); // Phase 3: Marketplace & Seller Discovery
  await registerCapabilityGatedRoutes(fastify, posRoutes, '/api/v1/pos', 'pos_sync'); // Disabled until POS certification evidence is recorded.
  await registerCapabilityGatedRoutes(fastify, deliveryRoutes, '/api/v1/delivery', 'delivery_partner'); // Disabled until delivery provider launch evidence is recorded.
  await fastify.register(couponRoutes, { prefix: '/api/v1/coupons' }); // Phase 3: Promotions & Coupons
  await registerCapabilityGatedRoutes(fastify, enhancedReferralRoutes, '/api/v1', 'enhanced_referrals_affiliates'); // Disabled until enhanced referral launch evidence is recorded.
  await registerCapabilityGatedRoutes(fastify, subscriptionRoutes, '/api/v1/subscriptions', 'subscriptions', { bypassPaths: ['/webhook'] }); // Billing lifecycle APIs disabled; signed Stripe webhooks remain verifiable.
  await fastify.register(whatsappRoutes, { prefix: '/api/v1/whatsapp' });
  await registerCapabilityGatedRoutes(fastify, ocrRoutes, '/api/v1/ocr', 'ocr_import'); // Disabled until OCR provider and privacy evidence are recorded.
  await fastify.register(referralRoutes, { prefix: '/api/v1/referrals' });
  await fastify.register(gdprRoutes, { prefix: '/api/v1/gdpr' });
  await fastify.register(reorderRoutes, { prefix: '/api/v1/reorder' });
  await fastify.register(adminRoutes, { prefix: '/api/v1/admin' }); // Phase 3: Admin Backend Platform
  await fastify.register(notificationRoutes, { prefix: '/api/v1/notifications' }); // iOS: Notifications
  await fastify.register(cartRoutes, { prefix: '/api/v1/cart' }); // iOS: Shopping cart
  await fastify.register(settingsRoutes, { prefix: '/api/v1/settings' }); // iOS: User settings

  // 404 handler
  fastify.setNotFoundHandler((request, reply) => {
    reply.status(404).send({
      success: false,
      error: {
        code: 'NOT_FOUND',
        message: `Route ${request.method} ${request.url} not found`,
      },
    });
  });
}

// Initialize database
async function initializeDatabase() {
  try {
    await AppDataSource.initialize();
    await assertNoPendingMigrationsForAppStartup(AppDataSource);
    fastify.log.info('Database connection established');
  } catch (error) {
    fastify.log.error({ err: error }, 'Failed to connect to database');
    throw error;
  }
}

// Start server
async function start() {
  try {
    // Initialize database
    await initializeDatabase();

    // Register plugins and routes
    await registerPlugins();
    await registerRoutes();

    // Set error handler
    fastify.setErrorHandler(errorHandler);

    // Start listening
    await fastify.listen({ port: PORT, host: HOST });

    fastify.log.info(`Server listening on http://${HOST}:${PORT}`);
    fastify.log.info(`API Documentation available at http://${HOST}:${PORT}/api/docs`);
    fastify.log.info(`Environment: ${process.env.NODE_ENV || 'development'}`);
  } catch (error) {
    fastify.log.error(error);
    process.exit(1);
  }
}

// Handle graceful shutdown
const gracefulShutdown = async (signal: string) => {
  fastify.log.info(`Received ${signal}, shutting down gracefully...`);
  try {
    await fastify.close();
    await AppDataSource.destroy();
    process.exit(0);
  } catch (error) {
    fastify.log.error({ err: error }, 'Error during shutdown');
    process.exit(1);
  }
};

process.on('SIGTERM', () => gracefulShutdown('SIGTERM'));
process.on('SIGINT', () => gracefulShutdown('SIGINT'));

// Start the server only when this module is the process entry point.
if (import.meta.url === `file://${process.argv[1]}`) {
  start();
}